move azure computeresource to separate package
