+++ azure/computeresource/computeresource_fake.go
@@ -1,4 +1,4 @@
-package controller
+package computeresource
 
 import (
 	"context"
